Convert span kind values to lowercase for Jaeger exporter
The OpenTelemetry Span.Kind enumerated type values are in uppercase, e.g. SERVER, CLIENT. These are incompatible with Jaeger which is based on the OpenTracing semantics where these are in lowercase.
The spans are collected as expected, but Jaeger is unable to display deep dependancy graphs as these are based on span kind values in lowercase, e.g. "server".