Add error tag to spans for Jaeger exporter
The auto-instrumentation libraries add errors to the spans when they occur. The error tag is not added to the spans by the Jaeger exporter if the span status is not OK.
When sending these spans to the OpenTelemetry Collector it will add the error tag when it translates them and sends them to the Jaeger collector. This process does not
take place when sending spans directly to a Jaeger collector (not using the OTel collector) and so the Jaeger UI does not display visual indicators for spans with errors.
Adding the error tag in the adapter if the span.status.code is not OK, and this tag hasn't already been set, means that it will be present when using either the OTel collector
or the Jaeger collector directly. The Otel collector will remove this tag automatically when it receives the span and sets the status code so this will not be sent as a duplicate.