Removed the worker thread
The worker thread has been removed and we now instead use the timer for scheduling work, thus saving us a thread. In addition, if there remains work to be done once the export has completed then we queue it for running immediately. A test has changed slightly in that it was previously relying on the worker thread becoming blocked; which it now no longer does.