Adjusted the stress test
Adjusted the stress test as it can only guarantee that the buffer size is sent to the exporter now. I'm noticing that several thousand get exported, but some do (and should) get lost. Also, the test did highlight one area of neglect where I should be waking up the timer when the buffer is half full and there's no export in progress.