Remove support to override create key
Indeed this removes one of the testing case otelAsGrpc, but we should not expose an API just for this purpose until we know for sure we need it (applying the rule of minimal API).

Also the createKey it is a strange method on the Storage (which should encapsulate only methods related to storage, and the creation of the key is a Context property).
In the future if needed this should be a method on the Context or other ContextKeyFactory, not on the Storage.

Signed-off-by: Bogdan Drutu <bogdandrutu@gmail.com>