Modify updateVersionInDocs task to run on all projects
This Gradle task updates the project version number found in the README. It previously only ran on the root README, leaving subprojects like extensions out of date.

This commit applies the task to all projects, and modifies the text replacement to be more flexible to accommodate the many ways of declaring dependencies in Gradle.