Remove unused package-info from context_prop io.otel.context (#1900)
The presence of package-info.java in the io.opentelemetry.context package of the io.opentelemetry.context.propagation JPMS module results in a split package, due to the io.opentelemetry.context module declaring the same package. This causes big problems for JPMS consumers.

Since the package-info.java is unused, removing it should not cause any problems.