Add custom proto marshaler to avoid unnecessary allocation
Some of the differences with the generated code:
1. Avoid builders to construct the marshalers;
2. Use less objects for Attributes because of the way `oneof` generated code is implemented; Also proto files supports arrays of any combination which is not the case for us.
3. Take advantage on the fact that data we know are immutable, no need to wrap to unmodifiable list/maps etc;

Unfortunately we cannot get the whole performance back because we need to write a custom Marshaler for grpc (which is possible but try to limit the PR size).

As the numbers show we have ~25% improvement in CPU and ~15% in memory with this PR. With a custom marshaler we can get anoter 7% CPU and 7% memory.

Signed-off-by: Bogdan Drutu <bogdandrutu@gmail.com>