Apply dep-mgmt also to unresolvable confs for composite build.
This should only make a difference when used in composite builds.
Without this, the dependency management versions won't be applied
when the ultimate user is a project outside this build. E.g. if
a project in a composite build uses
:odin-java:exporters:opentelemetry-exporter-zipkin,
it will indirectly depend on io.zipkin.reporter2:zipkin-reporter
without version.

Theoretically, this could be fixed in the using project by adding
a version constraint, but that is rather awkward.

With this change, including build usually work out of the box, but
they are forced to use a version compatible with the dependency
constraints from our dependency management.
This is especially annoying because we only have a single big dependencyManagement
configuration with production and test dependencies,
so e.g. our JUnit version constraint will be forced on the including build.