docs: Update the version in the README.md to algin with the latest release
During the release process the gradle task `updateVersionInDocs` should
be executed. This was previously part of the release process
documentation, but was removed.
https://github.com/open-telemetry/opentelemetry-java/commit/14e2fefe7af1290c7220fb44d31d73415e987d4f#diff-87d6a57a759f2386bf9caf5309b9aa0b9ff0ca70f5e38696e114a1ccdb0e785dL66

The gradle task is now part of the release workflow and performs an
update of the README.md after the maven upload.
The documentation changes will end up in a pull request together
with the apidiff baseline changes.